@@ -4,18 +4,18 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Utility
-  uuid: ^3.0.6
-  intl: ^0.18.1
+  uuid: ^4.4.2
+  intl: ^0.19.0
 
   # UI
-  flutter_credit_card: ^3.0.6
+  flutter_credit_card: ^4.0.1
 
   liqpay:
     path: ../
@@ -23,7 +23,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
 
 flutter:
   uses-material-design: true
