@@ -1,23 +1,20 @@
 name: liqpay
 description: LiqPay payment service wrapper library to be used on a client or server side dart.
-version: 0.0.1+3
+version: 0.0.2
 homepage: https://github.com/Newgarden-Solutions/liqpay
 repository: https://github.com/Newgarden-Solutions/liqpay
 issue_tracker: https://github.com/Newgarden-Solutions/liqpay/issues
 
 environment:
-  sdk: ">=3.0.5 <4.0.0"
+  sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
-  # Network
-  http: ^1.1.2
-  http_interceptor: ^2.0.0-beta.7
-
-  # Security
-  crypto: ^3.0.3
+  crypto: ^3.0.5
+  http: ^1.2.2
+  http_interceptor: ^2.0.0
 
 dev_dependencies:
-  lints: ^3.0.0
-  test: ^1.25.1
+  lints: ^4.0.0
+  test: ^1.25.8
   mockito: ^5.4.4
-  build_runner: ^2.4.8
+  build_runner: ^2.4.12
