@@ -7,7 +7,11 @@ enum LiqPayCurrency {
 
   const LiqPayCurrency(this.value);
 
-  static LiqPayCurrency fromValue(String value) {
+  static LiqPayCurrency? fromValue(String? value) {
+    if (value == null) {
+      return null;
+    }
+
     for (LiqPayCurrency entry in LiqPayCurrency.values) {
       if (entry.value == value) {
         return entry;
