@@ -1,13 +1,16 @@
 enum LiqPayLanguage {
   uk("uk"),
-  ru("ru"),
   en("en");
 
   final String value;
 
   const LiqPayLanguage(this.value);
 
-  static LiqPayLanguage fromValue(String value) {
+  static LiqPayLanguage? fromValue(String? value) {
+    if (value == null) {
+      return null;
+    }
+
     for (LiqPayLanguage entry in LiqPayLanguage.values) {
       if (entry.value == value) {
         return entry;
